# Support decoding into pre-allocated buffers