This is to better support re-using buffers (e.g. with buffer pools) for decode heavy workflows. 

I also have a staged change that automatically pools the `chunk` variable to automatically reduce allocations. If that is desirable, I can put in another PR for that.