@shanson7 2018-08-13
Benchmark to show the benefit of buffer reuse (obviously would vary in real world use):
```
go test -run=NONE -bench=.
goos: darwin
goarch: amd64
pkg: github.com/eapache/go-xerial-snappy
BenchmarkSnappyDecode-8             	 5000000	       255 ns/op	      80 B/op	       3 allocs/op
BenchmarkSnappyStreamDecode-8       	 1000000	      1276 ns/op	    3872 B/op	       7 allocs/op
BenchmarkSnappyDecodeInto-8         	10000000	       169 ns/op	       0 B/op	       0 allocs/op
BenchmarkSnappyStreamDecodeInto-8   	 2000000	       844 ns/op	    1088 B/op	       3 allocs/op
PASS
ok  	github.com/eapache/go-xerial-snappy	7.265s
```