@eapache 2018-08-13
Nice, thanks.

I'm not opposed to automatic pooling but unless the benefits are significant I'm not sure it's worth the complexity, especially since performance-sensitive workloads will want to use `DecodeInto` anyway?