@shanson7 2018-08-13
Well, `DecodeInto` still allocates the `chunk` array for performing `chunk` at a time decoding. Let me beef up the benchmark suite with some larger test data to see if there is a need.

Thanks!