@shanson7 2018-08-14
So, I determined that the buffer pool for the temp chunks is not worth it. 

It doesn't save any allocations, since the pool itself may allocate to store the chunk.
It saves a bit on *amount* allocated, but it's so very small compared to the size of the decoded object (one chunk size, of course).

Here's the full bench diff showing how little it changes:
```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkSnappyDecode-8                      270           301           +11.48%
BenchmarkSnappyDecodeInto-8                  200           194           -3.00%
BenchmarkSnappyStreamDecode-8                1626          1586          -2.46%
BenchmarkSnappyStreamDecodeInto-8            1041          969           -6.92%
BenchmarkSnappyStreamDecodeMassive-8         12978326      13355279      +2.90%
BenchmarkSnappyStreamDecodeIntoMassive-8     6065664       6197499       +2.17%

benchmark                                    old MB/s     new MB/s     speedup
BenchmarkSnappyDecode-8                      222.22       198.68       0.89x
BenchmarkSnappyDecodeInto-8                  299.20       309.07       1.03x
BenchmarkSnappyStreamDecode-8                657.93       674.32       1.02x
BenchmarkSnappyStreamDecodeInto-8            1027.13      1103.24      1.07x
BenchmarkSnappyStreamDecodeMassive-8         789.01       766.74       0.97x
BenchmarkSnappyStreamDecodeIntoMassive-8     1688.19      1652.28      0.98x

benchmark                                    old allocs     new allocs     delta
BenchmarkSnappyDecode-8                      3              3              +0.00%
BenchmarkSnappyDecodeInto-8                  0              0              +0.00%
BenchmarkSnappyStreamDecode-8                7              7              +0.00%
BenchmarkSnappyStreamDecodeInto-8            3              3              +0.00%
BenchmarkSnappyStreamDecodeMassive-8         16             18             +12.50%
BenchmarkSnappyStreamDecodeIntoMassive-8     1              1              +0.00%

benchmark                                    old bytes     new bytes     delta
BenchmarkSnappyDecode-8                      80            80            +0.00%
BenchmarkSnappyDecodeInto-8                  0             0             +0.00%
BenchmarkSnappyStreamDecode-8                3872          2881          -25.59%
BenchmarkSnappyStreamDecodeInto-8            1088          96            -91.18%
BenchmarkSnappyStreamDecodeMassive-8         54296576      54264872      -0.06%
BenchmarkSnappyStreamDecodeIntoMassive-8     58900         26164         -55.58%
```